Debugged a bunch of stuff, including reset.

+++ GPRO Net SDK/project/Unity/gpro-net-Unity/Assets/Scripts/AIScript.cs
@@ -36,7 +36,10 @@ public class AIScript : MonoBehaviour
     {
         if (collision.gameObject.CompareTag("Bullet") && isControlledLocally)
         {
-            client.DestroyLocalAI(id);
+            if (collision.GetComponent<BulletScript>().hasHitShield)
+            {
+                client.DestroyLocalAI(id);
+            }
         }
     }
